backend/pyproject.toml
[project]
name = "lineage-usage-agents-backend"
version = "0.1.0"
description = "Lineage and Usage agents — FastAPI backend"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "anthropic>=0.40",
    "google-cloud-storage>=2.19",
    "google-cloud-firestore>=2.19",
    "sqlglot>=25.30",
    "httpx>=0.28",
    "sse-starlette>=2.1",
    "python-multipart>=0.0.20",
    "orjson>=3.10",
    "tenacity>=9.0",
    "structlog>=24.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
    "mypy>=1.13",
]

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
